% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfgtools.R
\name{dfg_pack}
\alias{dfg_pack}
\title{Dynamic Feature Generator}
\usage{
dfg_pack(
  el,
  period = c("day", "week", "month", "quarter", "year", "sec", "min", "hour"),
  sequential = F,
  event_funs = c("count", "elapsed", "tte", "censored"),
  attr_funs = NULL,
  var_funs = NULL,
  horizon = NULL
)
}
\arguments{
\item{el}{(data.frame) The input event-log table.
The event-log must have these columns. Column names must match exactly: 
- \code{caseID, eventType, eventTime, attribute, value}
evenTime must be of class Date, otherwise, modify accordingly.}

\item{period}{(character) Choose period of time for which is periodic transition system is required.}

\item{sequential}{(logical) Should the dynamic features be generated for a sequence of periodic times.
If set to \code{FALSE}, the dynamic features are generated only for the times at which an event has happend.}

\item{event_funs}{(character): which types of features do you want to generate from the occurrence of events?}
}
\description{
This function generates a package of multiple tables containing dynamic features 
extracted from a given eventlog.
}
