% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_tools.R
\name{el2cp}
\alias{el2cp}
\title{Eventlog to Case-Profile Convertor}
\usage{
el2cp(eventlog, ...)
}
\description{
Directly converts eventlog to caseProfile bypassing caseVarProfile, 
but the aggregators must be a subset of:
sum, mean, last, first, min, max, count
}
