% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstools.R
\name{markovchain_transition_time_estimator}
\alias{markovchain_transition_time_estimator}
\title{Transition Time Generator for Markov-chanin Random Walk Simulation}
\usage{
markovchain_transition_time_estimator(
  histobj,
  input,
  current_time,
  family = c("normal", "exponential", "exp", "gaussian", "norm"),
  ...
)
}
\arguments{
\item{histobj}{object of class \code{TransitionSystem} containing history of all transitions upto the current time \code{current_time}}

\item{input}{dataframe containing transitions for which a transition end time needs to be generated}

\item{current_time}{POSIXct containing the current time. All transition end times will be after the specified current time.}
}
\value{
dataframe with additional column \code{pred_duration} in seconds.
}
\description{
This function generates random transition time (status durations) 
based on the average observed transition times
(given by argument \code{histobj}). 
MarkovChain transition time generator assumes transition times have normal distribution.
This is the default time generator engine for the transition system monte-carlo simulation.
}
