% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvis.R
\name{plot.process.tree}
\alias{plot.process.tree}
\title{plots process tree for the transition system}
\usage{
\method{plot}{process.tree}(obj, plotter = "sankeytree", ...)
}
\description{
plots process tree for the transition system
}
