% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvis.R
\name{plot_process_tree}
\alias{plot_process_tree}
\title{plots process tree for the transition system}
\usage{
plot_process_tree(obj, plotter = "sankeytree", config = NULL, ...)
}
\description{
plots process tree for the transition system
}
