% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvis.R
\name{plot_transition_time_histogram}
\alias{plot_transition_time_histogram}
\title{Shows the distribution of transition times (durations) in a histogram plot.}
\usage{
plot_transition_time_histogram(
  obj,
  source = NULL,
  target = NULL,
  remove_outliers = F,
  time_unit = c("second", "minute", "hour", "day", "week", "year"),
  plotter = c("plotly", "ggplot2"),
  ...
)
}
\arguments{
\item{obj}{object of class \code{TransitionSystem}: The Transition System object from which transition durations are extracted}

\item{source}{character: Transition source status. Must be within the source statuses of the transition system.}

\item{target}{character: Transition destination status. Must be within the target statuses of the transition system.
If you do not specify any of the two arguments \code{source} or \code{target}, 
a drop-down menu will apear where you can select your desired transition.
This feature is using package \code{crosstalk} and currently, 
only works when plotter is \code{plotly}.}

\item{remove_outliers}{logical: Should outliers be removed? 
If set to \code{TRUE}, 
outliers beyond four standard deviations from the mean, will be removed recursively.}

\item{time_unit}{character: time unit to be used in the output visualization. 
Can be one of these options: \code{'second', 'minute', 'hour', 'day', 'week', 'year'}.}

\item{plotter}{character: Which plotter package do you want to generate the visualization?
Currently, only these options are supported:
\code{'plotly', 'ggplot2'}}
}
\description{
Shows the distribution of transition times (durations) in a histogram plot.
}
