% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promer.R
\docType{package}
\name{promer}
\alias{promer}
\alias{promer-package}
\title{NIBE package for comprehensive process modelling in R}
\description{
This package has whatever you need to run a process mining analysis.
}
\section{Class TRANSYS}{

promer provides a Reference class named as TRANSYS.
}

