\name{rprom-package}
\alias{rprom-package}
\alias{rprom}
\docType{package}
\title{
\packageTitle{rprom}
}
\description{
\packageDescription{rprom}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rprom}
\packageIndices{rprom}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rprom}

Maintainer: \packageMaintainer{rprom}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
