% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\docType{class}
\name{Survival-class}
\alias{Survival-class}
\alias{Survival}
\title{Reference class for modelling survival in a process}
\description{
Reference class containing some properties and methods required for survival modelling.
}
\section{Fields}{

\describe{
\item{\code{data}}{list A list of tables. Table names can be: el (event log), cel (current eventlog), cp (case profile), ccp (current case profile), 
cvp (case-attribute profile), ccvp(current case-attribute profile), cpt (case profile time), ccpt (current case profile time), 
cpag (case profile aggregated), ccpag (current case profile aggregated)}

\item{\code{report}}{list A list of tables containing analysis reports.}

\item{\code{settings}}{list A list of class settings. Has following parameters: caseID_col, eventTime_col, eventType_col, attribute_col, value_col,
caseStart_tag, caseEnd_tag, age_varname, deathReason_varname, cp_attributes, normalize_hazard, cp_aggregators, cp_binners, cp_categoricals, cp_attributes}
}}

\section{Methods}{

\describe{
\item{\code{feed.caseProfile(dataset)}}{Feed case profile. Argument 'dataset' must be a sparklyr table. Column names must have been defined in settings}

\item{\code{feed.eventlog(dataset)}}{Feed eventlog. Argument 'dataset' must be a sparklyr table. Column names must have been defined in settings}

\item{\code{get.caseProfile(full = F)}}{Returns current values of case features}

\item{\code{get.caseProfileTime(full = T)}}{Returns timing properties of each case}

\item{\code{get.hazard(reasons = NULL, categoricals)}}{This function calculates hazard and survival probability from aggregated observations}

\item{\code{goto(time)}}{Takes you to a given time}

\item{\code{plot.death(
  reasons = NULL,
  categoricals = NULL,
  plotter = "plotly",
  gain = 100,
  t0 = 0
)}}{Plots Kaplan-Meier death chart}

\item{\code{plot.hazard(
  reasons = NULL,
  categoricals = NULL,
  plotter = "plotly",
  smooth_k = 6,
  gain = 100
)}}{Plots Kaplan-Meier hazard chart}

\item{\code{plot.survival(
  reasons = NULL,
  categoricals = NULL,
  plotter = "plotly",
  gain = 100,
  t0 = 0
)}}{Plots Kaplan-Meier survival chart}

\item{\code{reset(...)}}{Reset given table names in data list}

\item{\code{set(...)}}{Set a setting parameter to a value}
}}

