% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transys.R
\docType{class}
\name{TRANSYS-class}
\alias{TRANSYS-class}
\alias{TRANSYS}
\title{Reference class for modelling Transition Systems}
\description{
TRANSYS is a reference class containing some properties and methods required for analysing, modelling and visualising a Transition Sytem.
A Transistopn System is a Markov Chain model descibing a system which can change status over time.
}
\section{Fields}{

\describe{
\item{\code{history}}{data.frame holding history data of status transitions}
}}


