% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tstools.R
\name{markovchain_transition_classifier}
\alias{markovchain_transition_classifier}
\title{Generates predicted transition targets based on markov chain random walk simulation. 
This function is a simple transition probability model based on memory-less markov-chain model.
It randomly picks destination for each given transition respecting 
the probability distribution of each transition target.
Transition probability distribution are extracted from given argument \code{histobj}.
This is the default target generator engine for the transition system monte-carlo simulation.
The \code{TranSys} simulator engine calls this function multiple times during the sumulation.}
\usage{
markovchain_transition_classifier(histobj, input, ...)
}
\arguments{
\item{histobj}{object of class \code{TranSys} containing history of all transitions upto the current time \code{current_time}}

\item{input}{dataframe containing transitions for which a transition end time needs to be generated}

\item{current_time}{POSIXct containing the current time. All transition end times will be after the specified current time.}
}
\value{
dataframe with additional column \code{nextStatus}.
}
\description{
Generates predicted transition targets based on markov chain random walk simulation. 
This function is a simple transition probability model based on memory-less markov-chain model.
It randomly picks destination for each given transition respecting 
the probability distribution of each transition target.
Transition probability distribution are extracted from given argument \code{histobj}.
This is the default target generator engine for the transition system monte-carlo simulation.
The \code{TranSys} simulator engine calls this function multiple times during the sumulation.
}
